requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.0
torch>=2.0.0
transformers>=4.30.0
inseq>=0.5.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.15.0
requests>=2.25.0
beautifulsoup4>=4.11.0
Pillow>=9.0.0
markdown>=3.0.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
sentence-splitter>=1.0.0
thefuzz>=0.19.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0
networkx>=3.0
matplotlib>=3.6.0
seaborn>=0.12.0
tqdm>=4.65.0
datasets>=2.14.0
openai-whisper>=20230918
scipy>=1.10.0
